# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão de Colaboradores",
        "description": "Documentação da API de Gestão de Colaboradores utilizando Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Lista todas as empresas associadas ao usuário autenticado",
                "operationId": "54f966c056050f9db4fff6ee5a45a76a",
                "responses": {
                    "200": {
                        "description": "Lista de empresas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CompanyResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou não fornecido."
                    },
                    "403": {
                        "description": "O usuário não está associado a nenhuma empresa."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Criar uma nova empresa e associar um usuário admin",
                "operationId": "00e2db7b0f71c32b377f4dfc4dd685a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "cnpj",
                                    "opening_date",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Empresa ABC"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rua Exemplo, 123, São Paulo, SP"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "12345678000190"
                                    },
                                    "opening_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contato@empresaabc.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Empresa criada com sucesso."
                                        },
                                        "company": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Empresa ABC"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Rua Exemplo, 123, São Paulo, SP"
                                                },
                                                "cnpj": {
                                                    "type": "string",
                                                    "example": "12345678000190"
                                                },
                                                "opening_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-01-01"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "contato@empresaabc.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2023-01-01T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2023-01-01T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Erro ao associar o usuário autenticado como administrador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao associar o usuário ID X à empresa ID Y."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao associar o usuário à empresa."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar a empresa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao criar a empresa."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompanyResource": {
                "title": "Company Resource",
                "description": "Estrutura de resposta para informações de uma empresa.",
                "properties": {
                    "id": {
                        "description": "ID único da empresa.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome da empresa.",
                        "type": "string",
                        "example": "Empresa Exemplo LTDA"
                    },
                    "address": {
                        "description": "Endereço da empresa.",
                        "type": "string",
                        "example": "Rua Exemplo, 123, São Paulo, SP"
                    },
                    "cnpj": {
                        "description": "CNPJ da empresa.",
                        "type": "string",
                        "example": "12345678000195"
                    },
                    "opening_date": {
                        "description": "Data de abertura da empresa.",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "email": {
                        "description": "E-mail de contato da empresa.",
                        "type": "string",
                        "format": "email",
                        "example": "contato@empresaexemplo.com"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token no formato 'Bearer {seu_token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}